[Testsing] Adds test data files (#34)
Scope: introduces a test data files that are created alongside with code
files. Those files are pre-filled with sample tests from leetcode and
could be extended with the custom values. CLI test input has always a
priority over test files. If test file is missing and test input is not
provided, sample test case is used as was previously.

Changes:
- Fixes a bug with missing Vec<> inside of optional
- Adds runtime error message to the output
- Generates a test file when test is created