fix: Problems' description (pick, edit) (#163)
* fix: daily problem parsing

Changed a key to fix parsing daily problem request.
Fixed when id for daily problem should be fetched. Before, it will
always be fetched for the `pick` option regardless of the existence of
the `-d` option

* fix: comment problem desc

Changed desc comment to render Question.content instead of
Question.t_content, which was empty.